wapp_sentinel/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
requests>=2.26.0
openai>=0.27.0
APScheduler>=3.9.1
pydantic>=1.9.0
python-multipart>=0.0.5
aiohttp>=3.8.0
pytest>=7.0.0
httpx>=0.23.0  # For async HTTP requests
tenacity>=8.0.0  # For retries